test-mcp-server/requirements.txt
# Requirements for testing SSE endpoint with smolagents
smolagents
requests
sseclient-py
mcp